separated AssetShareButtons component
